Address and keystore related logic mixed together
Extract address related logic into separate module.
